+++ src/HL_nDimGaussian.cc
@@ -182,11 +182,11 @@ double HL_nDimGaussian::GetChi2(std::vector<double> theory)  //, double theory_e
   {
     throw std::runtime_error("Error in HL_nDimGaussian::GetChi2, you had different dimensions in theory and experiment");
   }
-  for(int i=0;i<central_restricted.size(); i++) { diff.push_back(central_restricted[i] - theory[i] );}
+  for(size_t i=0;i<central_restricted.size(); i++) { diff.push_back(central_restricted[i] - theory[i] );}
 
-  for (int i=0; i < HL_cov_inv_restricted.size1(); ++i)
+  for (size_t i=0; i < HL_cov_inv_restricted.size1(); ++i)
   {
-    for (int j=0; j<HL_cov_inv_restricted.size2(); ++j)
+    for (size_t j=0; j<HL_cov_inv_restricted.size2(); ++j)
     {
       chi2+= diff[i] * HL_cov_inv_restricted(i,j)*diff[j] ;
     }
@@ -250,12 +250,12 @@ double HL_nDimGaussian::GetChi2(std::vector<double> theory , boost::numeric::ubl
   {
     throw std::runtime_error("Error in HL_nDimGaussian::GetChi2, you had different dimensions in theory and experiment");
   }
-  for(int i=0;i<central_restricted.size(); i++) { diff.push_back(central_restricted[i] - theory[i] );}
+  for(size_t i=0;i<central_restricted.size(); i++) { diff.push_back(central_restricted[i] - theory[i] );}
 
 
-  for (int i=0; i < HL_cov_inv_restricted.size1(); ++i)
+  for (size_t i=0; i < HL_cov_inv_restricted.size1(); ++i)
   {
-    for (int j=0; j<HL_cov_inv_restricted.size2(); ++j)
+    for (size_t j=0; j<HL_cov_inv_restricted.size2(); ++j)
     {
       chi2+= diff[i] * HL_cov_inv_restricted(i,j)*diff[j] ;
     }
